Rearrainged documentation and the method structure
